support different approaches of patching used by peek-mysql2
