+++ lib/prepend_mysql2_patch.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module Rack
+  MINI_PROFILER_PREPEND_MYSQL2_PATCH = true
+end
